% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detect_package.R
\name{energy_anomaly_detect}
\alias{energy_anomaly_detect}
\title{anomaly_detect_package}
\usage{
energy_anomaly_detect(selected_data, weather_data, training_start_date,
  training_finish_date, test_start_date, test_finish_date)
}
\arguments{
\item{selected_data}{The data variable should be composed of the household number (id), date (date), and power usage (usage).}

\item{weather_data}{It is the data extracted by Korean weather (koreaweather) R package. The variables in the data are date, average temperature, relative humidity, and wind.}

\item{training_start_date}{Determine the start date of training data to use in setting the prediction model in the R package.}

\item{training_finish_date}{Determine the end date of the training data to use in setting the prediction model in the R package.}

\item{test_start_date}{Set the start date of test data to detect anomaly.}

\item{test_finish_date}{Set the finsih date of test data to detect anomaly.}
}
\description{
This package judges whether or not the power consumption per hour is abnomaly.
}
\examples{
 energy_anomaly_detect(example_household, weather_information,
 "2016-01-03 23:00:00","2016-01-21 23:00:00",
 "2016-01-22 00:00:00", "2016-01-22 05:00:00")
}
\keyword{anomaly}
\keyword{consumption}
\keyword{detection,}
\keyword{power}
